+++ src/vidhrdw/suprloco.c
@@ -202,7 +202,7 @@ static void render_sprite(struct osd_bitmap *bitmap,int spr_number)
 {
 	int sx,sy,col,row,height,src,adjy,dy;
 	unsigned char *spr_reg;
-	unsigned short *spr_palette;
+	UINT32 *spr_palette;
 	short skip;	/* bytes to skip before drawing each row (can be negative) */
