+++ src/vidhrdw/twin16.c
@@ -98,7 +98,7 @@ static void draw_text( struct osd_bitmap *bitmap ){
 static void draw_sprite( /* slow slow slow, but it's ok for now */
 	struct osd_bitmap *bitmap,
 	const UINT16 *pen_data,
-	const UINT16 *pal_data,
+	const UINT32 *pal_data,
 	int xpos, int ypos,
 	int width, int height,
 	int flipx, int flipy ){
@@ -233,7 +233,7 @@ static void draw_sprites( struct osd_bitmap *bitmap, int pri ){
 			int xpos = source[1];
 			int ypos = source[2];
 
-			const UINT16 *pal_data = Machine->pens+((attributes&0xf)+0x10)*16;
+			const UINT32 *pal_data = Machine->pens+((attributes&0xf)+0x10)*16;
 			int height	= 16<<((attributes>>6)&0x3);
 			int width	= 16<<((attributes>>4)&0x3);
 			const UINT16 *pen_data = 0;
@@ -373,7 +373,7 @@ static void draw_layer( struct osd_bitmap *bitmap, int opaque ){
 			*/
 			const data16_t *gfx_data = gfx_base + (code&0x7ff)*16 + bank_table[(code>>11)&0x3]*0x8000;
 			int color = (code>>13);
-			UINT16 *pal_data = Machine->pens + 16*(0x20+color+8*palette);
+			UINT32 *pal_data = Machine->pens + 16*(0x20+color+8*palette);
 
 			{
 				int y;
